requirements.txt
alembic==1.13.2
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
dash==2.17.1
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==3.0.3
Flask-Dance==7.1.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
flask-talisman==1.1.0
idna==3.7
importlib_metadata==8.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
nest-asyncio==1.6.0
numpy==2.0.0
oauthlib==3.2.2
packaging==24.1
pandas==2.2.2
plotly==5.22.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
retrying==1.3.4
setuptools==70.2.0
six==1.16.0
SQLAlchemy==2.0.31
tenacity==8.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
URLObject==2.4.3
Werkzeug==3.0.3
zipp==3.19.2
